python/aegis-ai/pyproject.toml
[project]
name = "aegis-ai"
version = "0.1.0"
description = "Tier 1 AI-assisted semantic security for AEGIS"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "python-dotenv",
    "httpx",
    "win11toast",
    "websockets"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# ── Pyrefly (IDE type-checker) configuration ──────────────────────────────────
# Tells Pyrefly where the real source root is so it can resolve all relative
# imports (from .llm, from .orchestrator.governance, etc.) correctly.
# Without this it guesses wrong and produces spurious virtual-buffer errors.
[tool.pyrefly]
project_includes = ["."]
search_path = ["."]
venv = ".venv"
python_version = "3.13"

# Per-file overrides: tests use sys.path injection so ignore missing src.*
[[tool.pyrefly.per_file_errors]]
path = "tests/**"
filter_errors = ["missing-import"]

# ── Pytest configuration ───────────────────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths  = ["tests"]
asyncio_mode = "auto"
